Revert "fix(frontend): やみモードの切り替え方法を変更"
