+++ packages/frontend/src/navbar.ts
@@ -9,11 +9,11 @@ import { clearCache } from './utility/clear-cache.js';
 import { $i } from '@/i.js';
 import { miLocalStorage } from '@/local-storage.js';
 import { openInstanceMenu, openToolsMenu } from '@/ui/_common_/common.js';
-import { changeNormalMode, changeYamiMode } from '@/change-mode.js';
 import { lookup } from '@/utility/lookup.js';
 import * as os from '@/os.js';
 import { i18n } from '@/i18n.js';
 import { unisonReload } from '@/utility/unison-reload.js';
+import { claimAchievement } from '@/utility/achievements.js';
 
 const roleManager = async () => {
 	try {
@@ -169,15 +169,104 @@ export const navbarItemDef = reactive({
 		},
 	},
 	mode: {
-		title: computed(() => $i?.isInYamiMode ? i18n.ts._yami._yamiModeSwitcher.yami : i18n.ts._yami._yamiModeSwitcher.normal),
+		title: i18n.ts._yami.switchMode,
 		icon: computed(() => $i?.isInYamiMode ? 'ti ti-moon' : 'ti ti-users-group'),
-		action: (ev) => {
-			if ($i?.isInYamiMode) {
-				changeNormalMode(ev);
-			}
-			if (!$i?.isInYamiMode) {
-				changeYamiMode(ev);
-			}
+		action: (ev: MouseEvent) => {
+			os.popupMenu([{
+				text: i18n.ts._yami._yamiModeSwitcher.normal,
+				active: $i && !$i.isInYamiMode,
+				action: async () => {
+					if ($i && $i.isInYamiMode) {
+						// 「今後表示しない」の設定を確認
+						const neverShowExitYamiModeInfo = miLocalStorage.getItem('neverShowExitYamiModeInfo');
+
+						// 表示しないが設定されていない場合はダイアログを表示
+						if (neverShowExitYamiModeInfo !== 'true') {
+							const confirm = await os.actions({
+								type: 'warning',
+								title: i18n.ts._yami.switchMode,
+								text: i18n.ts._yami._yamiModeSwitcher.exitYamiModeConfirm,
+								actions: [
+									{
+										value: 'yes' as const,
+										text: i18n.ts.ok,
+										primary: true,
+									},
+									{
+										value: 'neverShow' as const,
+										text: `${i18n.ts.ok} (${i18n.ts.neverShow})`,
+										danger: true,
+									},
+									{
+										value: 'cancel' as const,
+										text: i18n.ts.cancel,
+									},
+								],
+							});
+
+							if (confirm.canceled || confirm.result === 'cancel') return;
+
+							if (confirm.result === 'neverShow') {
+								miLocalStorage.setItem('neverShowExitYamiModeInfo', 'true');
+							}
+						}
+
+						os.apiWithDialog('i/update', {
+							isInYamiMode: false,
+						}).then(() => {
+							unisonReload();
+						});
+					}
+				},
+			}, {
+				text: i18n.ts._yami._yamiModeSwitcher.yami,
+				active: $i && $i.isInYamiMode,
+				action: async () => {
+					if ($i && !$i.isInYamiMode) {
+						// 「今後表示しない」の設定を確認
+						const neverShowEnterYamiModeInfo = miLocalStorage.getItem('neverShowEnterYamiModeInfo');
+
+						// 表示しないが設定されていない場合はダイアログを表示
+						if (neverShowEnterYamiModeInfo !== 'true') {
+							const confirm = await os.actions({
+								type: 'warning',
+								title: i18n.ts._yami.switchMode,
+								text: i18n.ts._yami._yamiModeSwitcher.enterYamiModeConfirm,
+								actions: [
+									{
+										value: 'yes' as const,
+										text: i18n.ts.ok,
+										primary: true,
+									},
+									{
+										value: 'neverShow' as const,
+										text: `${i18n.ts.ok} (${i18n.ts.neverShow})`,
+										danger: true,
+									},
+									{
+										value: 'cancel' as const,
+										text: i18n.ts.cancel,
+									},
+								],
+							});
+
+							if (confirm.canceled || confirm.result === 'cancel') return;
+
+							if (confirm.result === 'neverShow') {
+								miLocalStorage.setItem('neverShowEnterYamiModeInfo', 'true');
+							}
+						}
+
+						os.apiWithDialog('i/update', {
+							isInYamiMode: true,
+						}).then(() => {
+							unisonReload();
+							// やみモードに入った実績を解除
+							claimAchievement('markedAsYamiModeUser');
+						});
+					}
+				},
+			}], ev.currentTarget ?? ev.target);
 		},
 	},
 	about: {
@@ -197,7 +286,7 @@ export const navbarItemDef = reactive({
 	reload: {
 		title: i18n.ts.reload,
 		icon: 'ti ti-refresh',
-		action: () => {
+		action: (ev) => {
 			window.location.reload();
 		},
 	},
